+++ web-app/src/ng1/leaflet-extensions/FixedWidthMarker.js
@@ -1,18 +1,24 @@
 var L = require('leaflet');
 
 L.FixedWidthMarker = L.Marker.extend({
+  options: {
+    iconWidth: 35
+  },
 
   initialize: function(latlng, options) {
-    var self = this;
+    L.Marker.prototype.initialize.call(this, latlng, options);
+
+    const self = this;
     if (options.iconUrl) {
       options.icon = this.fixedWidthIcon = L.fixedWidthIcon({
         iconUrl: options.iconUrl,
         tooltip: options.tooltip,
+        iconWidth: this.options.iconWidth,
         onIconLoad: function() {
           if (self._popup && self._icon) {
-            var iconHeight = $(self._icon).height();
+            const iconHeight = $(self._icon).height();
 
-            var popup = self.getPopup();
+            const popup = self.getPopup();
             popup.options.offset = [0, (iconHeight - 8) * -1];
             popup.update();
           }