+++ web-app/src/ng1/leaflet-extensions/index.js
@@ -1,6 +1,7 @@
 require('./GridLayer');
 require('./FixedWidthIcon');
 require('./FixedWidthMarker');
+require('./ObservationMarker');
 require('./LocationMarker');
 require('./LocationMarker.Popup');
 require('./WorldExtentControl');